Split ``sphinx.ext.intersphinx._resolve``
